+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router-dom';
 import Menu from './components/Menu';
-//import Formation from './components/Formation/Formation';
+import Formation from './components/Formation/Formation';
 import Characters from './components/Characters/Characters';
 import './App.css';
 
@@ -10,7 +10,7 @@ class App extends Component {
 		return (
 			<div className="App">
 				<Route exact={true} path="/" component={Menu} />
-				
+				<Route path="/formation/" component={Formation} />
 				<Route path="/characters/" component={Characters} />
 			</div>
 		);